+++ CHANGELOG.rst
@@ -13,6 +13,8 @@ Breaking API Changes
 * finagle-core: Correct the spelling of `Tracing.recordClientSendFrargmet()` to
   `Tracing.recordClientSendFragment()` ``PHAB_ID=D505617``
 
+* finagle-redis: Use `StrictKeyCommand` for XDEL ``PHAB_ID=D517291``
+
 * finagle-zipkin-scribe: The scribe.thrift file was moved to finagle-thrift/src/main/thrift under a new
   namespace. `com.twitter.finagle.thrift.scribe.(thrift|thriftscala)` ``PHAB_ID=D511471``